restructuring of frontend folders
