+++ frontend/cg/cg.component.scss
@@ -1,4 +1,4 @@
-@import 'base';
+@import '../base';
 
 body {
   margin: 0;
