+++ frontend/cg/cg.module.ts
@@ -4,14 +4,14 @@ import { FormsModule } from "@angular/forms";
 import { HttpModule } from "@angular/http";
 
 import { CrowdgezwitscherComponent } from "./cg.component";
-import { MaterialModule } from "./material.module";
+import { MaterialModule } from "../material.module";
 import { RoutingModule } from "./cg.routing";
 
-import { MapComponent } from "./map.component";
-import { ContactComponent } from "./contact.component";
+import { MapComponent } from "../map/map.component";
+import { ContactComponent } from "../contact/contact.component";
 
-import { MapService } from "./map.service";
-import { ContactService } from "./contact.service";
+import { MapService } from "../map/map.service";
+import { ContactService } from "../contact/contact.service";
 
 @NgModule({
     imports: [
