+++ frontend/contact/contact.component.scss
@@ -8,6 +8,7 @@
 
 .contact-card-error {
   background-color: #f2dede;
+  white-space: pre-line;
 
   .md-card {
     display: flex;
