+++ frontend/main.ts
@@ -4,7 +4,7 @@
 import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
 
 import { Config } from "./config/config";
-import { CrowdgezwitscherModule } from "./cg.module";
+import { CrowdgezwitscherModule } from "./cg/cg.module";
 
 let config = new Config();
 config.configure();
