+++ frontend/map/map.component.scss
