+++ frontend/map/mapObject.ts
